+++ src/point.rs
@@ -3,15 +3,15 @@ use std::ops;
 
 #[derive(Copy, Clone, Debug)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 impl Point {
-    pub const W: f32 = 1.0;
+    pub const W: f64 = 1.0;
 
-    pub fn new(x: f32, y: f32, z: f32) -> Self {
+    pub fn new(x: f64, y: f64, z: f64) -> Self {
         Self { x, y, z }
     }
 }
