+++ src/ray.rs
@@ -16,7 +16,7 @@ impl Ray {
         Self { origin, direction }
     }
 
-    pub fn position(self, t: f32) -> Point {
+    pub fn position(self, t: f64) -> Point {
         self.origin + self.direction * t
     }
